Don't stop dev until the analytics event is sent

+++ packages/app/src/cli/services/dev/ui/components/DevSessionUI.tsx
@@ -18,6 +18,7 @@ import {openURL} from '@shopify/cli-kit/node/system'
 import figures from '@shopify/cli-kit/node/figures'
 import {isUnitTest} from '@shopify/cli-kit/node/context/local'
 import {treeKill} from '@shopify/cli-kit/node/tree-kill'
+import {postRunHookHasCompleted} from '@shopify/cli-kit/node/hooks/postrun'
 import {Writable} from 'stream'
 
 interface DevSesionUIProps {
@@ -61,9 +62,17 @@ const DevSessionUI: FunctionComponent<DevSesionUIProps> = ({
       await onAbort()
     }
     if (isUnitTest()) return
-    treeKill(process.pid, 'SIGINT', false, () => {
-      process.exit(0)
-    })
+
+    // Wait for the post run hook to complete or timeout after 5 seconds.
+    let totalTime = 0
+    setInterval(() => {
+      if (postRunHookHasCompleted() || totalTime > 5000) {
+        treeKill(process.pid, 'SIGINT', false, () => {
+          process.exit(0)
+        })
+      }
+      totalTime += 100
+    }, 100)
   })
 
   const errorHandledProcesses = useMemo(() => {
